Small refactor trying to put type alias for transactional event handler
